Refactor component to manage local storage
